+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.2.0
+
+* Renamed pages API endpoint.
+* Added pageGroup directive, which returns a list of pages filtered by a group.
+
 # 1.1.1
 
 * Bump ADRF requirement.